+++ R/bayesian.R
@@ -214,14 +214,6 @@ update.bayesian <-
            silent = NULL,
            fresh = FALSE,
            ...) {
-    if (!exists("update_spec", where = "package:parsnip", mode = "function")) {
-      parsnip::update_dot_check(...)
-
-      if (!is.null(parameters)) {
-        parameters <- parsnip::check_final_param(parameters)
-      }
-    }
-
     args <- list(
       formula.override = rlang::enquo(formula.override),
       family = rlang::enquo(family),
@@ -251,46 +243,14 @@ update.bayesian <-
       silent = rlang::enquo(silent)
     )
 
-    if (exists("update_spec", where = "package:parsnip", mode = "function")) {
-      update_spec <- utils::getFromNamespace("update_spec", "parsnip")
-      update_spec(
-        object = object,
-        parameters = parameters,
-        args_enquo_list = args,
-        fresh = fresh,
-        cls = "bayesian",
-        ...
-      )
-    } else {
-      args <- parsnip::update_main_parameters(args, parameters)
-
-      eng_args <- parsnip::update_engine_parameters(object$eng_args, fresh = fresh, ...)
-
-      if (fresh) {
-        object$args <- args
-        object$eng_args <- eng_args
-      } else {
-        null_args <- purrr::map_lgl(args, parsnip::null_value)
-        if (any(null_args)) {
-          args <- args[!null_args]
-        }
-        if (length(args) > 0) {
-          object$args[names(args)] <- args
-        }
-        if (length(eng_args) > 0) {
-          object$eng_args[names(eng_args)] <- eng_args
-        }
-      }
-
-      parsnip::new_model_spec(
-        "bayesian",
-        args = object$args,
-        eng_args = object$eng_args,
-        mode = object$mode,
-        method = NULL,
-        engine = object$engine
-      )
-    }
+    parsnip::update_spec(
+      object = object,
+      parameters = parameters,
+      args_enquo_list = args,
+      fresh = fresh,
+      cls = "bayesian",
+      ...
+    )
   }
 
 # -------------------------------------------------------------------------