+++ R/bayesian_make.R
@@ -28,11 +28,7 @@ bayesian_make <- function(modes = c("classification", "regression")) {
     # -------------------------------------------------------------------------
 
     for (pkg in dependpkgs) {
-      if ("mode" %in% rlang::fn_fmls_names(parsnip::set_dependency)) {
-        parsnip::set_dependency(model, engine, pkg = pkg, mode = mode)
-      } else {
-        parsnip::set_dependency(model, engine, pkg = pkg)
-      }
+      parsnip::set_dependency(model, engine, pkg = pkg, mode = mode)
     }
 
     # -------------------------------------------------------------------------